% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregions.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Terrestrial Ecoregions Of the World (TEOW).}
\format{
A SpatialPolygonDataFrame with 847 features and 11 variables:
\describe{
\item{OBJECTID}{numeric identifier of the geographic feature}
\item{ECO_NAME}{name of the ecoregion}
\item{BIOME_NUM}{numeric identifier of the biome}
\item{BIOME_NAME}{name of the biome}
\item{REALM}{name of the realm}
\item{ECO_BIOME}{alphanumeric identifier of the eco-biome}
\item{NNH}{...}
\item{ECO_ID}{numeric identifier of the ecoregion}
\item{SHAPE_LENG}{length of the perimeter of the ecoregion, in kilometers}
\item{SHAPE_AREA}{area of the ecoregion, in square kilometers}
\item{LICENSE}{name of the Creative Commons attribution of the data}

}
}
\source{
\url{https://ecoregions2017.appspot.com/}
}
\usage{
ecoregions
}
\description{
A revised map of the TEOW (Dinerstein, 2017).
}
\seealso{
\url{https://academic.oup.com/bioscience/article/67/6/534/3102935}, \url{https://academic.oup.com/bioscience/article/51/11/933/227116}
}
\keyword{datasets}
