% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGeodata.R
\docType{package}
\name{RGeodata}
\alias{RGeodata}
\title{RGeodata: Global geographic data useful for spatial analysis}
\description{
The package gives users access to global geographic data useful for spatial analysis in biogeography,
conservation biology and any fields requiring the use of such data.
}
\details{
The current version contains a revised map of Terrestrial Ecoregions Of the World (TEOW) by (Dinerstein, 2017), and
a world map of terrestrial lands derived from TEOW.
Details about \pkg{RGeodata}
}
\seealso{
\url{https://academic.oup.com/bioscience/article/67/6/534/3102935}, \url{https://academic.oup.com/bioscience/article/51/11/933/227116}
}
