% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terrestrial_lands.R
\docType{data}
\name{terrestrial_lands}
\alias{terrestrial_lands}
\title{Terrestrial Lands Of the World (TLOW).}
\format{
A Simple Features Collections with 144749 features and 6 fields. The fields have no special meaning (except \code{area}) but are by-products of geoprocessing:
\describe{
\item{id}{...}
\item{level}{...}
\item{source}{...}
\item{parent_id}{...}
\item{sibling_id}{...}
\item{area}{area of the territory, in square kilometers}
}
}
\source{
\url{https://ecoregions2017.appspot.com/}
}
\usage{
terrestrial_lands
}
\description{
A map of the Terrestrial lands of the world, derived from Ecoregions Of the World (Dinerstein, 2017) with boundaries between adjacent ecoregions units dissolved (eliminated).
}
\seealso{
\url{https://academic.oup.com/bioscience/article/67/6/534/3102935}, \url{https://academic.oup.com/bioscience/article/51/11/933/227116}
}
\keyword{datasets}
