% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregions.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Terrestrial Ecoregions Of the World (TEOW).}
\format{
A Simple Features Collections with 847 features and 5 fields:
\describe{
\item{ECO_NAME}{name of the ecoregion}
\item{BIOME_NUM}{numeric identifier of the biome}
\item{BIOME_NAME}{name of the biome}
\item{REALM}{name of the realm}
\item{ECO_BIOME}{alphanumeric identifier of the eco-biome}

}
}
\source{
\url{https://ecoregions2017.appspot.com/}
}
\usage{
ecoregions
}
\description{
A revised map of the TEOW (Dinerstein, 2017).
}
\keyword{datasets}
